Fixed bbReplace function

+++ BlitzSharp3D/bbruntime/bbstring.cpp
@@ -25,7 +25,7 @@ BBStr* bbRight(BBStr* s, int n) {
 
 BBStr* bbReplace(BBStr* s, BBStr* from, BBStr* to) {
 	int n = 0, from_sz = from->size(), to_sz = to->size();
-	while (n < (int)(s->size() && (n = s->find(*from, n)) != string::npos)) {
+	while (n < s->size() && (n = s->find(*from, n)) != string::npos) {
 		s->replace(n, from_sz, *to);
 		n += to_sz;
 	}
